Some changes to tests and more comments
