+++ src/Stratis.Bitcoin.Tests/Consensus/ChainedHeaderTreeTest.cs
@@ -152,67 +152,4 @@ namespace Stratis.Bitcoin.Tests.Consensus
             Assert.Equal(listOfNewHeaders.First(), connectedHeadersNew.DownloadFrom.Header);
         }
     }
-
-    /// TODO: move reflection methods to a shared test utils project (as soon as there is one) that is referenced by all test projects. 
-    public static class ReflectionExtensions
-    {
-        public static Dictionary<uint256, HashSet<int>> GetPeerIdsByTipHash(this ChainedHeaderTree chainedHeaderTree)
-        {
-            return chainedHeaderTree.GetMemberValue("peerIdsByTipHash") as Dictionary<uint256, HashSet<int>>;
-        }
-
-        public static Dictionary<int, uint256> GetPeerTipsByPeerId(this ChainedHeaderTree chainedHeaderTree)
-        {
-            return chainedHeaderTree.GetMemberValue("peerTipsByPeerId") as Dictionary<int, uint256>;
-        }
-
-        public static Dictionary<uint256, ChainedHeader> GetChainedHeadersByHash(this ChainedHeaderTree chainedHeaderTree)
-        {
-            return chainedHeaderTree.GetMemberValue("chainedHeadersByHash") as Dictionary<uint256, ChainedHeader>;
-        }
-
-        public static object GetMemberValue(this object obj, string memberName)
-        {
-            MemberInfo memberInfo = GetMemberInfo(obj, memberName);
-
-            if (memberInfo == null)
-                throw new Exception("memberName");
-
-            if (memberInfo is PropertyInfo)
-                return memberInfo.As<PropertyInfo>().GetValue(obj, null);
-
-            if (memberInfo is FieldInfo)
-                return memberInfo.As<FieldInfo>().GetValue(obj);
-
-            throw new Exception();
-        }
-
-        private static MemberInfo GetMemberInfo(object obj, string memberName)
-        {
-            var propertyInfos = new List<PropertyInfo>();
-
-            propertyInfos.Add(obj.GetType().GetProperty(memberName, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.FlattenHierarchy));
-            propertyInfos = Enumerable.ToList(Enumerable.Where(propertyInfos, i => !ReferenceEquals(i, null)));
-            if (propertyInfos.Count != 0)
-                return propertyInfos[0];
-
-            var fieldInfos = new List<FieldInfo>();
-
-            fieldInfos.Add(obj.GetType().GetField(memberName, BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.FlattenHierarchy));
-
-            // To add more types of properties.
-            fieldInfos = Enumerable.ToList(Enumerable.Where(fieldInfos, i => !ReferenceEquals(i, null)));
-
-            if (fieldInfos.Count != 0)
-                return fieldInfos[0];
-
-            return null;
-        }
-
-        [System.Diagnostics.DebuggerHidden]
-        private static T As<T>(this object obj)
-        {
-            return (T)obj;
-        }
-    }
 }
